#1015 Move tag_usage method in application_controller.rb into its own helper
