+++ app/controllers/application_controller.rb
@@ -15,6 +15,7 @@ class ApplicationController < ActionController::Base
   include ActionView::Helpers::DateHelper
   include ActionView::Helpers::TextHelper
   include CommentsHelper
+  include TagsHelper
 
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
@@ -143,29 +144,7 @@ class ApplicationController < ActionController::Base
 
     data.empty? ? {} : top_three_focus(data)
   end
-  # rubocop:enable MethodLength
-
-  # rubocop:disable MethodLength
-  def tag_usage(data_id, data_type, user_id)
-    result = []
-    moments = user_moments(user_id).order('created_at DESC')
-    if data_type == 'category'
-      strategies = user_strategies(user_id).order('created_at DESC')
-      [moments, strategies].each do |records|
-        objs = []
-        records.find_each do |r|
-          objs.push(r) if data_included?(data_type, data_id, r)
-        end
-        result << objs
-      end
-    elsif data_type.in?(%w[mood strategy])
-      moments.find_each do |m|
-        result << m if data_included?(data_type, data_id, m)
-      end
-    end
-    result
-  end
-  # rubocop:enable MethodLength
+  # rubocop:enable MethodLength  
 
   # rubocop:disable MethodLength
   # TODO: move this logic out of the controller and into a helper method
@@ -198,12 +177,6 @@ class ApplicationController < ActionController::Base
   # rubocop:enable MethodLength
 
   private
-
-  def data_included?(data_type, data_id, data)
-    return false unless data_type.in?(%w[category mood strategy])
-
-    data_id.in?(data[data_type])
-  end
 
   # TODO: refactor calling method to pass a hash to start with
   def top_three_focus(data)
